añadido subida de fotos

+++ resources/views/artesania/artesania.blade.php
@@ -11,15 +11,19 @@
     @foreach($artesanias as $artesania)
       <div class="col-md-4 col-sm-6">
         {!! $artesania->post !!}
-        <a class="btn btn-primary" href="{{ route('artesanias.edit', $artesania) }}">Editar</a>
-        {!! Form::open(['route' => ['artesanias.destroy', $artesania], 'method' => 'DELETE']) !!}
-          {!! Form::submit('Borrar', ['class' => 'btn btn-danger']) !!}
-        {!! Form::close() !!}
+        @if (Auth::check())
+          <a class="btn btn-primary" href="{{ route('artesanias.edit', $artesania) }}">Editar</a>
+          {!! Form::open(['route' => ['artesanias.destroy', $artesania], 'method' => 'DELETE']) !!}
+            {!! Form::submit('Borrar', ['class' => 'btn btn-danger']) !!}
+          {!! Form::close() !!}
+        @endif
       </div>
     @endforeach
   </div>
-  <div class="cold-md-12">
-    <a class="btn btn-primary" href="{{ route('artesanias.create') }}">Crea un post nuevo</a>
-  </div>
+  @if (Auth::check())
+    <div class="cold-md-12">
+      <a class="btn btn-primary" href="{{ route('artesanias.create') }}">Crea un post nuevo</a>
+    </div>
+  @endif
 </div>
 @endsection
